requirements.txt
atari-py
gym
matplotlib
numpy
opencv-python
Pillow
tensorflow-gpu==2.0.0a0
tqdm